# Improve Pedersen hash performance