Implemented alongside @gtank!

We first started by implementing w-ary non-adjacent form exponentiation, which shaved about 40% of the time off running Pedersen hash (for the merkle tree). We tried also to improve the doubling performance a little bit by using the specific doubling formula rather than the generic addition formula.

Finally, we just implemented fixed window tables and shaved about 90% of the cost off.

Here are some benchmarks:

```
at first:

test bench_pedersen_hash ... bench:     452,241 ns/iter (+/- 24,567)

after wnaf:

test bench_pedersen_hash ... bench:     276,877 ns/iter (+/- 7,185)

after more optimal doubling (eyeroll at the variance):

test bench_pedersen_hash ... bench:     279,210 ns/iter (+/- 7,150)

after dedicated window tables:

test bench_pedersen_hash ... bench:      37,373 ns/iter (+/- 2,020)
```