@daira 2018-07-26
BTW, might as well squash out the WNAF implementation (I just reviewed the overall diff, so didn't include that in my review).